remove: MemberServiceImpl 삭제 #15
